Added UI for displaying timer messages.
Created an empty game object to hold a script TimerManager to host all
the logic to controll the timer logic.
Once the player reaches point b a timer is triggered and they need to
get back to point a before it expires.
Created Canvas object with some text childs to host the UI elements.