Refactor package organization
Realized that the gitignore was assuming the base directory for the
project was the root directory but unity created the project on a
subfolder.
Moved all the filed up from the original unity folder. This resulted on
a lot of file churn but it reduced the amount of files in the repo.