Refactor MapManager to use “Grid” tiles
Instead of using a hardcoded charmap we will have an in editor tilemap
creator that we will use to spawn tiles and then customize by hand.
All the generated new tiles will exists under an empty game object
called Grid which will be used by MapManager to load the level.